SVN + Sonar data

+++ src/main/java/org/neo/gomina/api/projects/ProjectDetail.java
@@ -15,7 +15,7 @@ public class ProjectDetail {
     public String latest;
     public String released;
 
-    public Integer loc;
+    public Double loc;
     public Double coverage;
 
     public List<CommitLogEntry> commitLog;
